Use generated api types in vasu
