+++ frontend/src/employee-frontend/components/vasu/components/FollowupQuestion.tsx
@@ -9,7 +9,7 @@ import { useTranslation } from 'employee-frontend/state/i18n'
 import { UserContext } from '../../../state/user'
 import TextArea from 'lib-components/atoms/form/TextArea'
 import { Dimmed, H2, Label } from 'lib-components/typography'
-import { Followup, FollowupEntry } from '../vasu-content'
+import { Followup, FollowupEntry } from 'lib-common/api-types/vasu'
 import { QuestionProps } from './question-props'
 import QuestionInfo from '../QuestionInfo'
 import { VasuTranslations } from 'lib-customizations/employee'
