+++ frontend/src/employee-frontend/components/vasu/components/MultiSelectQuestion.tsx
@@ -7,7 +7,7 @@ import Checkbox from 'lib-components/atoms/form/Checkbox'
 import { FixedSpaceColumn } from 'lib-components/layout/flex-helpers'
 import { Label } from 'lib-components/typography'
 import { Gap } from 'lib-components/white-space'
-import { MultiSelectQuestion, QuestionOption } from '../vasu-content'
+import { MultiSelectQuestion, QuestionOption } from 'lib-common/api-types/vasu'
 import { ValueOrNoRecord } from './ValueOrNoRecord'
 import { QuestionProps } from './question-props'
 import QuestionInfo from '../QuestionInfo'
