+++ frontend/src/employee-frontend/components/vasu/components/RadioGroupQuestion.tsx
@@ -7,7 +7,7 @@ import Radio from 'lib-components/atoms/form/Radio'
 import { FixedSpaceColumn } from 'lib-components/layout/flex-helpers'
 import { Label } from 'lib-components/typography'
 import { Gap } from 'lib-components/white-space'
-import { QuestionOption, RadioGroupQuestion } from '../vasu-content'
+import { QuestionOption, RadioGroupQuestion } from 'lib-common/api-types/vasu'
 import { ValueOrNoRecord } from './ValueOrNoRecord'
 import QuestionInfo from '../QuestionInfo'
 import { VasuTranslations } from 'lib-customizations/employee'
