+++ frontend/src/employee-frontend/components/vasu/components/TextQuestion.tsx
@@ -6,7 +6,7 @@ import React from 'react'
 import InputField from 'lib-components/atoms/form/InputField'
 import TextArea from 'lib-components/atoms/form/TextArea'
 import { Label } from 'lib-components/typography'
-import { TextQuestion } from '../vasu-content'
+import { TextQuestion } from 'lib-common/api-types/vasu'
 import { ValueOrNoRecord } from './ValueOrNoRecord'
 import { QuestionProps } from './question-props'
 import QuestionInfo from '../QuestionInfo'
