+++ frontend/src/employee-frontend/components/vasu/sections/EvaluationDiscussionSection.tsx
@@ -6,7 +6,7 @@ import React, { Dispatch, SetStateAction } from 'react'
 import styled from 'styled-components'
 import { ContentArea } from 'lib-components/layout/Container'
 import { H2, H3, Label } from 'lib-components/typography'
-import { EvaluationDiscussionContent } from '../api'
+import { EvaluationDiscussionContent } from 'lib-common/generated/api-types/vasu'
 import TextArea from 'lib-components/atoms/form/TextArea'
 import { defaultMargins, Gap } from 'lib-components/white-space'
 import { DatePickerClearableDeprecated } from 'lib-components/molecules/DatePickerDeprecated'
