+++ frontend/src/employee-frontend/components/vasu/templates/CopyTemplateModal.tsx
@@ -6,12 +6,13 @@ import { FixedSpaceColumn } from 'lib-components/layout/flex-helpers'
 import { Label } from 'lib-components/typography'
 import React, { useState } from 'react'
 import FiniteDateRange from 'lib-common/finite-date-range'
+import { VasuTemplateSummary } from 'lib-common/generated/api-types/vasu'
 import { UUID } from 'lib-common/types'
 import InputField from 'lib-components/atoms/form/InputField'
 import { DatePickerDeprecated } from 'lib-components/molecules/DatePickerDeprecated'
 import FormModal from 'lib-components/molecules/modals/FormModal'
 import { useTranslation } from '../../../state/i18n'
-import { copyVasuTemplate, VasuTemplateSummary } from './api'
+import { copyVasuTemplate } from './api'
 
 interface Props {
   template: VasuTemplateSummary
